smartguard-service/requirements.txt
# SmartGuard ETMS — behavioral anomaly detection microservice
torch>=2.0.0
numpy>=1.24.0
paho-mqtt>=2.0.0
pyyaml>=6.0
